chore: proper default handling removing necessity for most of the nulls
