+++ src/ElectronBackend/main/user-settings.ts
@@ -7,17 +7,20 @@ import log from 'electron-log';
 import settings from 'electron-settings';
 
 import { AllowedFrontendChannels } from '../../shared/ipc-channels';
+import { DEFAULT_USER_SETTINGS } from '../../shared/shared-constants';
 import { UserSettings as IUserSettings } from '../../shared/shared-types';
-
-const DEFAULT_USER_SETTINGS = {
-  showClassifications: true,
-};
 
 export class UserSettings {
   public static async init() {
     if (process.argv.includes('--reset') || process.env.RESET) {
       log.info('Resetting user settings');
-      await settings.set(DEFAULT_USER_SETTINGS);
+      await settings.set(DEFAULT_USER_SETTINGS as unknown as never);
+    } else {
+      const currentSettings = await settings.get();
+      await settings.set({
+        ...DEFAULT_USER_SETTINGS,
+        ...currentSettings,
+      });
     }
   }
