+++ src/Frontend/Components/AttributionPanels/AttributionPanels.tsx
@@ -5,15 +5,13 @@
 import { useCallback } from 'react';
 
 import { AllowedFrontendChannels } from '../../../shared/ipc-channels';
+import { DEFAULT_PANEL_SIZES } from '../../../shared/shared-constants';
 import { text } from '../../../shared/text';
 import {
   useFilteredAttributions,
   useFilteredSignals,
 } from '../../state/variables/use-filtered-data';
-import {
-  DEFAULT_PANEL_SIZES,
-  usePanelSizes,
-} from '../../state/variables/use-panel-sizes';
+import { usePanelSizes } from '../../state/variables/use-panel-sizes';
 import { ResizePanels } from '../ResizePanels/ResizePanels';
 import { Container } from './AttributionPanels.style';
 import { AttributionsPanel } from './AttributionsPanel/AttributionsPanel';
